fix: event query 파일구조 변경
