+++ src/api/axios/Home/homeInterface.ts
@@ -1,0 +1,7 @@
+export interface EventData {
+  id: number;
+  image: string;
+  name: string;
+  description: string;
+  period: string;
+}
